// model/audit.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class Audit {
    /**
    * Subject identifier
    */
    'subject'?: string;
    /**
    * Subject name
    */
    'name'?: string;
    /**
    * Subject type
    */
    'type'?: string;
    /**
    * Actor performing the action
    */
    'actor'?: string;
    /**
    * Action performed
    */
    'action'?: string;
    /**
    * Additional audit information
    */
    'extras'?: { [key: string]: string; };
    /**
    * Audit entry creation timestamp
    */
    'createdAt'?: Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "subject",
            "baseName": "subject",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "actor",
            "baseName": "actor",
            "type": "string"
        },
        {
            "name": "action",
            "baseName": "action",
            "type": "string"
        },
        {
            "name": "extras",
            "baseName": "extras",
            "type": "{ [key: string]: string; }"
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return Audit.attributeTypeMap;
    }
}