// model/authGroup.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthGroupMember } from './authGroupMember';

export class AuthGroup {
    /**
    * The unique identifier of the approval group
    */
    'id'?: string;
    /**
    * The name of the approval group
    */
    'name': string;
    /**
    * The threshold for the approval group
    */
    'threshold': AuthGroup.ThresholdEnum;
    'description'?: string;
    'users'?: Array<AuthGroupMember>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "threshold",
            "baseName": "threshold",
            "type": "AuthGroup.ThresholdEnum"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "users",
            "baseName": "users",
            "type": "Array<AuthGroupMember>"
        }    ];

    static getAttributeTypeMap() {
        return AuthGroup.attributeTypeMap;
    }
}

export namespace AuthGroup {
    export enum ThresholdEnum {
        _1 = <any> '1',
        _2 = <any> '2',
        _3 = <any> '3',
        _4 = <any> '4',
        _5 = <any> '5',
        Any = <any> 'any',
        All = <any> 'all'
    }
}