// model/authGroupMember.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AuthGroupMember {
    /**
    * The unique identifier of the user
    */
    'id'?: string;
    /**
    * The email address of the user
    */
    'email'?: string;
    /**
    * The first name of the user
    */
    'firstName'?: string;
    /**
    * The last name of the user
    */
    'lastName'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "firstName",
            "baseName": "firstName",
            "type": "string"
        },
        {
            "name": "lastName",
            "baseName": "lastName",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthGroupMember.attributeTypeMap;
    }
}