// model/authorizeReply.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthorizeReplyOneOf } from './authorizeReplyOneOf';
import { AuthorizeReplyOneOf1 } from './authorizeReplyOneOf1';

export class AuthorizeReply {
    /**
    * Transaction ID for the transaction that\'s been sent to a user
    */
    'txID'?: string;
    /**
    * Approval ID for the approval that\'s been sent to an approval group
    */
    'approvalID'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "txID",
            "baseName": "txID",
            "type": "string"
        },
        {
            "name": "approvalID",
            "baseName": "approvalID",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthorizeReply.attributeTypeMap;
    }
}