// model/authorizeReplyOneOf.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Transaction response
*/
export class AuthorizeReplyOneOf {
    /**
    * Transaction ID for the transaction that\'s been sent to a user
    */
    'txID'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "txID",
            "baseName": "txID",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthorizeReplyOneOf.attributeTypeMap;
    }
}