// model/authorizeReplyOneOf1.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Approval response
*/
export class AuthorizeReplyOneOf1 {
    /**
    * Approval ID for the approval that\'s been sent to an approval group
    */
    'approvalID'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "approvalID",
            "baseName": "approvalID",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthorizeReplyOneOf1.attributeTypeMap;
    }
}