// model/checkTransactionStatus200Response.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CheckTransactionStatus200Response {
    'state'?: CheckTransactionStatus200Response.StateEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "state",
            "baseName": "state",
            "type": "CheckTransactionStatus200Response.StateEnum"
        }    ];

    static getAttributeTypeMap() {
        return CheckTransactionStatus200Response.attributeTypeMap;
    }
}

export namespace CheckTransactionStatus200Response {
    export enum StateEnum {
        Pending = <any> 'pending',
        Verified = <any> 'verified',
        Denied = <any> 'denied',
        Cancelled = <any> 'cancelled',
        Expired = <any> 'expired'
    }
}