// model/connectWise.ts
/**
 * SendAuth API
 * Authentication and identity management API for SendAuth platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ConnectWise {
    /**
    * Whether ConnectWise integration is enabled
    */
    'enabled'?: boolean;
    /**
    * ConnectWise server URL
    */
    'serverURL'?: string;
    /**
    * ConnectWise company ID
    */
    'companyID'?: string;
    /**
    * ConnectWise public key
    */
    'publicKey'?: string;
    /**
    * ConnectWise private key
    */
    'privateKey'?: string;
    /**
    * ConnectWise API version
    */
    'apiVersion'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "enabled",
            "baseName": "enabled",
            "type": "boolean"
        },
        {
            "name": "serverURL",
            "baseName": "serverURL",
            "type": "string"
        },
        {
            "name": "companyID",
            "baseName": "companyID",
            "type": "string"
        },
        {
            "name": "publicKey",
            "baseName": "publicKey",
            "type": "string"
        },
        {
            "name": "privateKey",
            "baseName": "privateKey",
            "type": "string"
        },
        {
            "name": "apiVersion",
            "baseName": "apiVersion",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ConnectWise.attributeTypeMap;
    }
}